feat(systemd): add a condition on WAYLAND_DISPLAY
The service would only start if the condition is meet. There is already
some logic checking this at startup in src/backend/mod.rs:174, but with
this the service is not even started, saving resuorces and time.

The aforementioned existing check continues to make sense since it is
an extra check just to be sure.